+++ daiv/automation/agents/plan_and_execute/conf.py
@@ -1,7 +1,6 @@
 from pydantic import Field
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
-from automation.agents.base import ThinkingLevel
 from automation.agents.constants import ModelName
 
 
@@ -16,17 +15,11 @@ class PlanAndExecuteSettings(BaseSettings):
     PLANNING_MODEL_NAME: ModelName = Field(
         default=ModelName.CLAUDE_3_7_SONNET, description="Model name to be used to plan tasks."
     )
-    PLANNING_THINKING_LEVEL: ThinkingLevel = Field(
-        default=ThinkingLevel.MEDIUM, description="Thinking level to be used to plan tasks."
-    )
     EXECUTION_MODEL_NAME: ModelName = Field(
         default=ModelName.CLAUDE_3_7_SONNET, description="Model name to be used to execute tasks."
     )
-    EXECUTION_THINKING_LEVEL: ThinkingLevel = Field(
-        default=ThinkingLevel.MEDIUM, description="Thinking level to be used to execute tasks."
-    )
     PLAN_APPROVAL_MODEL_NAME: ModelName = Field(
-        default=ModelName.GPT_4O_MINI, description="Model name to be used to evaluate the plan approval."
+        default=ModelName.GEMINI_2_0_FLASH, description="Model name to be used to evaluate the plan approval."
     )