+++ daiv/automation/agents/snippet_replacer/agent.py
@@ -53,7 +53,7 @@ class SnippetReplacerAgent(BaseAgent[Runnable[SnippetReplacerInput, SnippetRepla
         """
         if settings.STRATEGY == "llm" and self.validate_max_token_not_exceeded(input_data):
             return self._prompt | self.get_model(model=settings.MODEL_NAME).with_structured_output(
-                SnippetReplacerOutput
+                SnippetReplacerOutput, method="function_calling"
             )
         return RunnableLambda(self._replace_content_snippet)